// projects/ui-builder-component-lib-test/src/app/enums/routes.enum.ts
export enum RoutePaths {
  HomeComponent = '',
  BottomModalDemoComponent = 'bottom-modal',
  ButtonsDemoComponent = 'buttons',
  CircleProgressComponent = 'circle-progress',
  ColumnManagementModalDemoComponent = 'column-management-modal',
  ContentViewCardDemoComponent = 'content-view-card',
  DashedCheckboxTestComponent = 'dashed-checkbox',
  ExpansionPanelDemoComponent = 'expansion-panel',
  Feedback = 'feedback',
  FormFieldDemoComponent = 'form-field',
  FormGeneratorDemoComponent = 'form-generator',
  FormGeneratorModelsComponent = 'form-generator-models',
  GridDemoComponent = 'grid',
  ImageCardTestComponent = 'image-card',
  InfoCardDemoComponent = 'info-card',
  InfoCardListDemoComponent = 'info-card-list',
  InfoCardPlaceholderDemoComponent = 'info-card-placeholder',
  InputDateFieldTestComponent = 'input-date-field',
  InputFieldSelectTestComponent = 'input-field-select',
  InputFieldTestComponent = 'input-field',
  InputTimeFieldTestComponent = 'input-time-field',
  JsonEditorDemoComponent = 'json-editor',
  ListDemoComponent = 'list',
  LoginInputDemoComponent = 'login-input',
  MapDemoComponent = 'map',
  MenuDemoComponent = 'menu',
  MenuItemDemoComponent = 'menu-item',
  MultiFieldDemoComponent = 'multi-field',
  MultiselectDemoComponent = 'multiselect',
  NotificationBannerDemoComponent = 'notification-banner',
  NotificationsListDemoComponent = 'notifications-list',
  OtpInputTestComponent = 'otp-input',
  PercentageBarComponent = 'percentage-bar',
  PhotosUploadDemoComponent = 'photo-upload',
  PopperDemoComponent = 'popper',
  PopupModalDemoComponent = 'popup-modal',
  PremiumChipDemoComponent = 'premium-chip',
  ProfileCardComponent = 'profile-card',
  ProfileHeaderComponent = 'profile-header',
  ProgressbarDemoComponent = 'progressbar',
  RangeSelectorDemoComponent = 'range-selector',
  RatingFieldDemoComponent = 'rating-field',
  SearchBarDemoComponent = 'search-bar',
  SectionSeparatorDemoComponent = 'section-separator',
  SelectableDemoComponent = 'selectable',
  SelectOptionDemoComponent = 'select-option',
  TabMenuDemoComponent = 'tab-menu',
  TelInputDemoComponent = 'tel-input',
  TextBoxTestComponent = 'text-box',
  TextListDemoComponent = 'text-list',
  ToastNotificationDemoComponent = 'toast',
  ToastrDemoComponent = 'toastr',
  ToggleDemoComponent = 'toggle',
  TooltipDemoComponent = 'tooltip',
  TypeaheadFieldDemoComponent = 'typeahead-field',
}